+++ bin/save_api_data_to_db.py
@@ -1,14 +1,14 @@
 """ Script will run periodically and retrieve basic player data that remains unchanged. """
 import time
-import player_data
+import player_manager
 from requests import exceptions
 
 while True:
     try:
         print("Looking for Player Data to save:")
-        player_data.save_player_data()
+        player_manager.save_player_data(refresh_from_api=True)
         print("Looking for Event Data to save:")
-        player_data.save_event_data()
+        player_manager.save_event_data(refresh_from_api=True)
         SECONDS = 0
         for i in range(SECONDS):
             print("", end="\r")