+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Add script for adding a new player to the DB by name
+
 ## [0.3.3] - 2021-04-30
 ### Changed
 - Cleaned up the home page so it doesn't look absolutely horrible